move extension install logic to base component
